Добавить сохранение локейшна в адресной строчке при перезагрузке или заходе по ссылке
