+++ client/src/components/App/AppBehavior.tsx
@@ -1,14 +1,17 @@
 import { Provider } from 'react-redux'
-import { configureStore } from '../../redux/authToolkitRedux'
+import { BrowserRouter as Router } from 'react-router-dom'
 
+import { configureStore } from '../../redux/authToolkitRedux'
 import { AppTemplate } from './AppTemplate'
 
 export const store = configureStore()
 
-export function AppBehavior() {
+export const AppBehavior = () => {
   return (
     <Provider store={store}>
-      <AppTemplate />
+      <Router>
+        <AppTemplate />
+      </Router>
     </Provider>
   )
 }
